Load OpenSSH RSA keys without ed25519 dependencies
OpenSSH private key format is a container format that can hold RSA,
ECDSA, and ED25519 keys. KeyFactory previously required ED25519 support
for any OpenSSH-format private key before checking the contained key type.

Move the OpenSSH private key loader out of the ED25519 namespace and use
it directly from KeyFactory. Keep the old ED25519::OpenSSHPrivateKeyLoader
constant as a compatibility alias.

This allows OpenSSH-format RSA/ECDSA private keys to be loaded without
the ed25519 and bcrypt_pbkdf dependencies.
